requirements.txt
opencv-python==4.10.0.84
numba==0.60.0
numpy==1.26.4
tqdm==4.66.4
Pillow==10.4.0
matplotlib==3.9.0
torch==2.3.1+cu118
torchvision==0.18.1+cu118
tensorboardX==2.6.2.2
